requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
pydantic-settings
psycopg[binary]
psycopg-pool
openai
orjson
beautifulsoup4
lxml
passlib[bcrypt]
python-jose[cryptography]